+++ ICD.Connect.Partitioning.Commercial/Properties/AssemblyInfo.cs
@@ -1,0 +1,7 @@
+﻿using System.Reflection;
+
+[assembly: AssemblyTitle("ICD.Connect.Partitioning.Commercial")]
+[assembly: AssemblyCompany("ICD Systems")]
+[assembly: AssemblyProduct("ICD.Connect.Partitioning.Commercial")]
+[assembly: AssemblyCopyright("Copyright © ICD Systems 2019")]
+[assembly: AssemblyVersion("10.5.0.0")]