+++ ICD.Connect.Partitioning/Rooms/AbstractRoomSettings.cs
@@ -4,7 +4,6 @@ using ICD.Common.Utils;
 using ICD.Common.Utils.Extensions;
 using ICD.Common.Utils.Xml;
 using ICD.Connect.Settings;
-using ICD.Connect.Settings.Attributes.SettingsProperties;
 
 namespace ICD.Connect.Partitioning.Rooms
 {
@@ -15,7 +14,6 @@ namespace ICD.Connect.Partitioning.Rooms
 	public abstract class AbstractRoomSettings : AbstractSettings, IRoomSettings
 	{
 		private const string COMBINE_PRIORITY_ELEMENT = "CombinePriority";
-		private const string DIALINGPLAN_ELEMENT = "DialingPlan";
 
 		private const string PANELS_ELEMENT = "Panels";
 		private const string PANEL_ELEMENT = "Panel";
@@ -51,9 +49,6 @@ namespace ICD.Connect.Partitioning.Rooms
 		#region Properties
 
 		public int CombinePriority { get; set; }
-
-		[PathSettingsProperty("DialingPlans", ".xml")]
-		public string DialingPlan { get; set; }
 
 		public Dictionary<int, eCombineMode> Devices { get { return m_Devices; } }
 		public Dictionary<int, eCombineMode> Ports { get { return m_Ports; } }
@@ -93,8 +88,6 @@ namespace ICD.Connect.Partitioning.Rooms
 
 			writer.WriteElementString(COMBINE_PRIORITY_ELEMENT, IcdXmlConvert.ToString(CombinePriority));
 
-			writer.WriteElementString(DIALINGPLAN_ELEMENT, DialingPlan);
-
 			WriteChildrenToXml(writer, m_Panels, PANELS_ELEMENT, PANEL_ELEMENT);
 			WriteChildrenToXml(writer, m_Ports, PORTS_ELEMENT, PORT_ELEMENT);
 			WriteChildrenToXml(writer, m_Devices, DEVICES_ELEMENT, DEVICE_ELEMENT);
@@ -115,8 +108,6 @@ namespace ICD.Connect.Partitioning.Rooms
 			base.ParseXml(xml);
 
 			CombinePriority = XmlUtils.TryReadChildElementContentAsInt(xml, COMBINE_PRIORITY_ELEMENT) ?? 0;
-
-			DialingPlan = XmlUtils.TryReadChildElementContentAsString(xml, DIALINGPLAN_ELEMENT);
 
 			IEnumerable<KeyValuePair<int, eCombineMode>> panels = ReadListFromXml(xml, PANELS_ELEMENT, PANEL_ELEMENT);
 			IEnumerable<KeyValuePair<int, eCombineMode>> ports = ReadListFromXml(xml, PORTS_ELEMENT, PORT_ELEMENT);