+++ CHANGELOG.md
@@ -46,6 +46,19 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
  - DialingPlan configuration is now a single path
  - Added VolumePoints and Partitions to the room console
 
+## [10.5.0] - 2019-10-01
+### Added
+ - Added wake and sleep console commands to commercial room
+
+### Changed
+ - Moved commercial room abstractions and interfaces into Rooms subdirectory
+
+## [10.4.0] - 2019-09-17
+### Added
+ - Added commercial room abstractions and interfaces
+ - Added WakeSchedule features
+ - Added DialingPlanInfo
+
 ## [10.3.0] - 2019-08-02
 ### Added
  - Added console command to print child rooms for a given room